[NEW] favicons, componentize sections and elements, new mobile sidemenu
