+++ src/assets/styles/scss/_variables.scss
@@ -9,3 +9,5 @@ $font-weight-light: 300;
 $font-weight-regular: 400;
 $font-weight-medium: 500;
 $font-weight-bold: 700;
+
+$desktop-min-width: 600px;
