+++ src/assets/styles/scss/settings.scss
@@ -1,13 +1,18 @@
 html {
+  height: 100vh;
+  width: 100vw;
   overflow: hidden;
   scroll-behavior: smooth;
 }
 
 body {
+  overflow: hidden;
   height: 100%;
+  width: 100%;
 
   #app {
     overflow: hidden;
     height: 100%;
+    width: 100%;
   }
 }
