+++ src/locales/en-US.json
@@ -27,7 +27,7 @@
       "sections": {
         "intro": {
           "title": "Home",
-          "role": "Mobile and Full stack Developer",
+          "role": "Mobile and Full Stack Developer",
           "download_cv": "Download CV"
         },
         "projects": {
