+++ src/locales/pt-BR.json
@@ -27,7 +27,7 @@
       "sections": {
         "intro": {
           "title": "Início",
-          "role": "Desenvolvedor Mobile e Full stack",
+          "role": "Desenvolvedor Mobile e Full Stack",
           "download_cv": "Baixar CV"
         },
         "projects": {
