+++ src/views/PortfolioView.vue
@@ -1,38 +1,18 @@
 <template>
   <div
+    ref="content"
     class="main-container"
-    ref="introSection"
   >
-    <div
-      ref="navBar"
-      :class="['navbar-container', { hidden: currentSectionIndex == 0 }]"
-    >
-      <nav class="nav-bar">
-        <a
-          href="/"
-          class="logo"
-          :data-text="t('names.jp_kenji')"
-          >{{ t('names.jp_kenji') }}</a
-        >
-        <ul>
-          <li
-            v-for="section in sections"
-            :key="section.index"
-            :class="{ active: currentSectionIndex == section.index }"
-          >
-            <a
-              class="nav-item"
-              :href="section.navBarHref"
-              @click.prevent="scrollToSection(section.index)"
-            >
-              {{ section.title }}
-            </a>
-          </li>
-        </ul>
-        <a class="logo">{{ t('names.jp_kenji') }}</a>
-      </nav>
-      <div class="horizontal-line"></div>
-      <div class="vertical-line"></div>
+    <MainNavBar
+      :sections="sections"
+      :current-section-index="currentSectionIndex"
+      @scroll-to-section="scrollToSection"
+    />
+    <div :class="['mobile-menu-button', { hidden: !showMobileHeader }]">
+      <v-icon
+        icon="mdi-menu"
+        @click="openMobileMenu"
+      />
     </div>
     <div
       :class="[
@@ -56,167 +36,19 @@
       </div>
     </div>
     <div class="content">
-      <section
-        id="intro"
-        class="present"
-      >
-        <div class="profile-container">
-          <div class="profile">
-            <div class="name-container">
-              <span
-                class="jp-name"
-                :data-text="t('names.jp_kenji')"
-                >{{ t('names.jp_kenji') }}</span
-              >
-              <div
-                ref="sortLettersName"
-                :data-value="`${t('names.author')}*${t('names.author_jp')}`"
-                class="pt-name"
-              ></div>
-            </div>
-            <div class="role">
-              {{ t('features.portfolio.sections.intro.role') }}
-            </div>
-            <TechnologiesList ref="technologiesList" />
-            <a
-              href="/src/assets/files/Curriculo_Nicolas.pdf"
-              target="_blank"
-              rel="noopener noreferrer"
-              class="cv-link"
-              >{{ t('features.portfolio.sections.intro.download_cv') }}</a
-            >
-          </div>
-        </div>
-        <!-- Descomentar quando tiver certeza do conteúdo a mostrar aqui -->
-        <!-- <div class="languages-showcase-container"></div> -->
-      </section>
-      <section
+      <IntroSection
+        ref="introSection"
+        :is-mobile="mobile"
+        :social-links="socialLinks"
+      />
+      <ProjectsSection
         ref="projectsSection"
-        id="projects"
-      >
-        <SectionTitle :jp-text="KATAKANA_PROJECT">{{
-          t('features.portfolio.sections.projects.title')
-        }}</SectionTitle>
-        <div class="description">
-          <span>{{
-            t('features.portfolio.sections.projects.description')
-          }}</span>
-        </div>
-        <div class="projects-container">
-          <div
-            class="projects-list"
-            ref="projectsList"
-          >
-            <div
-              class="project"
-              v-for="project in projects"
-              :key="project.id"
-            >
-              <ProjectItem
-                @click="openProjectUrl(project.url)"
-                :project="project"
-              />
-            </div>
-          </div>
-          <div
-            class="previous-button hidden"
-            ref="projectsListPreviousButton"
-          >
-            <v-icon icon="mdi-chevron-left"></v-icon>
-          </div>
-          <div
-            class="next-button"
-            ref="projectsListNextButton"
-          >
-            <v-icon icon="mdi-chevron-right"></v-icon>
-          </div>
-        </div>
-      </section>
-      <section
-        id="contact"
+        :social-links="socialLinks"
+      />
+      <ContactSection
         ref="contactSection"
-      >
-        <div class="separator"></div>
-        <div class="contact-container">
-          <SectionTitle :jp-text="KATAKANA_CONTACT">{{
-            t('features.portfolio.sections.contact.title')
-          }}</SectionTitle>
-          <div class="description">
-            {{ t('features.portfolio.sections.contact.description') }}
-          </div>
-          <div class="email-contact-container">
-            <a
-              :href="`mailto:${MY_EMAIL}`"
-              class="email-contact"
-              >{{ MY_EMAIL }}</a
-            >
-            <div
-              ref="clipboardButton"
-              class="clipboard-button"
-            >
-              <ClipboardCheck
-                class="clipboard-checked"
-                v-if="clipboardShowCheck"
-              />
-              <Clipboard
-                class="clipboard-plain"
-                v-else
-              />
-            </div>
-          </div>
-          <div class="social-links-container">
-            <a
-              v-for="socialLink in socialLinks"
-              :key="socialLink.label"
-              :aria-label="socialLink.label"
-              :href="socialLink.url"
-              class="social-link"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              <component :is="socialLink.icon" />
-            </a>
-          </div>
-        </div>
-        <footer>
-          <div class="social-links-container">
-            <a
-              v-for="socialLink in socialLinks"
-              :key="socialLink.label"
-              :aria-label="socialLink.label"
-              :href="socialLink.url"
-              class="social-link"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              <component :is="socialLink.icon" />
-            </a>
-          </div>
-          <span class="label">{{
-            t('features.portfolio.sections.contact.footer.description')
-          }}</span>
-          <div class="repo-tip">
-            <span>{{
-              t(
-                'features.portfolio.sections.contact.footer.repo_tip.first_part',
-              )
-            }}</span>
-            <a
-              :href="GITHUB_REPO_URL"
-              target="_blank"
-              rel="noopener noreferrer"
-              >{{
-                t('features.portfolio.sections.contact.footer.repo_tip.link')
-              }}</a
-            >
-            <span>{{
-              t(
-                'features.portfolio.sections.contact.footer.repo_tip.second_part',
-              )
-            }}</span>
-          </div>
-        </footer>
-      </section>
+        :social-links="socialLinks"
+      />
       <div class="tip-footer">
         <div
           :class="['scroll-tip-container', { gone: currentSectionIndex != 0 }]"
@@ -260,52 +92,86 @@
       <div class="theme-selector"></div>
     </div>
   </div>
+  <div :class="['mobile-menu-container', { hidden: !isMobileMenuOpen }]">
+    <div class="close-button">
+      <v-icon
+        @click="closeMobileMenu"
+        icon="mdi-close"
+      />
+    </div>
+    <div class="navigation">
+      <a
+        href="/"
+        class="logo"
+        :data-text="t('names.jp_kenji')"
+        >{{ t('names.jp_kenji') }}</a
+      >
+      <ul>
+        <li
+          v-for="section in sections"
+          :key="section.index"
+          :class="[{ active: currentSectionIndex == section.index }]"
+        >
+          <a
+            class="nav-item"
+            :href="section.navBarHref"
+            @click.prevent="mobileNavSection(section.index)"
+          >
+            {{ section.title }}
+          </a>
+        </li>
+      </ul>
+    </div>
+    <div class="social-links-container">
+      <div
+        class="social-info"
+        style="flex-direction: row; justify-content: center"
+      >
+        <a
+          v-for="socialLink in socialLinks"
+          :key="socialLink.label"
+          :aria-label="socialLink.label"
+          :href="socialLink.url"
+          class="social-link"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          <component :is="socialLink.icon" />
+        </a>
+      </div>
+    </div>
+  </div>
 </template>
 
 <script lang="ts" setup>
   import Github from '@/components/icons/Github.vue';
   import Linkedin from '@/components/icons/Linkedin.vue';
-  import Clipboard from '@/components/icons/Clipboard.vue';
-  import ClipboardCheck from '@/components/icons/ClipboardCheck.vue';
   import ScrollTip from '@/components/icons/ScrollTip.vue';
+  import IntroSection from '@/components/sections/IntroSection.vue';
+  import ProjectsSection from '@/components/sections/ProjectsSection.vue';
+  import ContactSection from '@/components/sections/ContactSection.vue';
   import LocaleSelector from '@/components/LocaleSelector.vue';
-  import ProjectItem from '@/components/ProjectItem.vue';
-  import SectionTitle from '@/components/SectionTitle.vue';
-  import TechnologiesList from '@/components/TechnologiesList.vue';
-  import {
-    GITHUB_URL,
-    GITHUB_REPO_URL,
-    KATAKANA_CONTACT,
-    KATAKANA_PROJECT,
-    LINKEDIN_URL,
-    MY_EMAIL,
-  } from '@/helpers/constants';
-  import { projectsMock } from '@/helpers/projectsMock';
+  import { GITHUB_URL, LINKEDIN_URL, MY_EMAIL } from '@/helpers/constants';
   import { sortLettersAnimation } from '@/helpers/sortLettersAnimation';
-  import { useSnackbarStore } from '@/stores/snackbar_store';
   import { computed, onMounted, ref, watch } from 'vue';
   import type { Component, Ref } from 'vue';
 
   import { useI18n } from 'vue-i18n';
-  const { t } = useI18n();
+  import MainNavBar from '@/components/MainNavBar.vue';
+  import { useDisplay } from 'vuetify';
 
-  const snackbarStore = useSnackbarStore();
+  const { t } = useI18n();
+  const { mobile } = useDisplay({ mobileBreakpoint: 600 });
 
   const currentSectionIndex = ref(0);
-  const projects = ref(projectsMock);
-  const clipboardShowCheck = ref(false);
-
-  const navBar = ref();
-  const sortLettersName = ref();
-  const projectsList = ref();
-  const projectsListPreviousButton = ref();
-  const projectsListNextButton = ref();
-  const technologiesList = ref();
-  const clipboardButton = ref();
 
   const introSection = ref();
   const projectsSection = ref();
   const contactSection = ref();
+  const content = ref();
+
+  const isMobileMenuOpen = ref(false);
+  const showMobileHeader = ref(false);
 
   const socialLinks: ISocialLink[] = [
     {
@@ -326,7 +192,7 @@
       index: 0,
       title: t('features.portfolio.sections.intro.title'),
       ref: introSection,
-      navBarHref: '#',
+      navBarHref: '#intro',
     },
     {
       id: 'projects',
@@ -344,10 +210,6 @@
     },
   ]);
 
-  const openProjectUrl = (url: string) => {
-    window.open(url, '_blank', 'noopener,noreferrer');
-  };
-
   const scrollToSection = (newSectionIndex: number) => {
     const sectionInList = sections.value.find(
       section => section.index == newSectionIndex,
@@ -355,22 +217,17 @@
 
     if (!sectionInList) return;
 
+    const section = document.getElementById(
+      `${sectionInList.navBarHref}`.replace('#', ''),
+    ) as HTMLElement;
+
     currentSectionIndex.value = newSectionIndex;
     presentSections();
-    sectionInList.ref.value.scrollIntoView({ behavior: 'smooth' });
-  };
-
-  const startMainNameAnimation = () => {
-    const element = sortLettersName.value as HTMLElement;
-
-    if (!element) return;
-
-    let originalText = element.dataset.value || '';
-    const names = originalText.split('*');
-
-    originalText = element.textContent == names[0] ? names[1] : names[0];
-
-    sortLettersAnimation(element, { original: originalText });
+    if (section) {
+      section.scrollIntoView({
+        behavior: 'smooth',
+      });
+    }
   };
 
   const presentSections = () => {
@@ -385,24 +242,7 @@
 
   const setupInitialAnimations = () => {
     setTimeout(() => {
-      const nameContainer = document.querySelector('.name-container');
-
-      if (nameContainer) {
-        setTimeout(() => {
-          nameContainer.classList.add('present');
-        }, 2000);
-      }
-
-      startMainNameAnimation();
       presentSections();
-      setTimeout(() => {
-        technologiesList.value.startAnimation();
-      }, 1000);
-
-      sortLettersName.value.addEventListener(
-        'mouseover',
-        startMainNameAnimation,
-      );
 
       document
         .querySelectorAll('.sort-letters')
@@ -410,6 +250,7 @@
           sortLettersAnimation(element, {
             speedMs: 30,
           });
+
           element.addEventListener('mouseover', event =>
             sortLettersAnimation(event.target as HTMLElement, {
               speedMs: 30,
@@ -427,52 +268,9 @@
     }
   };
 
-  const setupProjectsList = () => {
-    projectsList.value.addEventListener('scroll', event => {
-      const scrollLeft = projectsList.value.scrollLeft;
-      const scrollWidth = projectsList.value.scrollWidth;
-      const clientWidth = projectsList.value.clientWidth;
-
-      if (Math.ceil(scrollLeft + clientWidth) >= scrollWidth) {
-        projectsListNextButton.value.classList.add('hidden');
-      } else {
-        projectsListNextButton.value.classList.remove('hidden');
-      }
-
-      if (scrollLeft <= 0) {
-        projectsListPreviousButton.value.classList.add('hidden');
-      } else {
-        projectsListPreviousButton.value.classList.remove('hidden');
-      }
-    });
-
-    projectsListNextButton.value.addEventListener('click', () => {
-      const scrollLeft = projectsList.value.scrollLeft;
-      const scrollWidth = projectsList.value.scrollWidth;
-      const clientWidth = projectsList.value.clientWidth;
-
-      if (Math.ceil(scrollLeft + clientWidth) >= scrollWidth) return;
-
-      projectsList.value.scrollTo({
-        left: scrollLeft + clientWidth / 2,
-        behavior: 'smooth',
-      });
-    });
-
-    projectsListPreviousButton.value.addEventListener('click', () => {
-      const scrollLeft = projectsList.value.scrollLeft;
-      const clientWidth = projectsList.value.clientWidth;
-
-      if (scrollLeft <= 0) return;
-
-      projectsList.value.scrollTo({
-        left: scrollLeft - clientWidth / 2,
-        behavior: 'smooth',
-      });
-    });
-  };
-
   const handleMainScroll = (event: WheelEvent) => {
+    if (mobile.value) return;
+
     const nextSection = sections.value[currentSectionIndex.value + 1];
     const previousSection = sections.value[currentSectionIndex.value - 1];
 
@@ -485,15 +283,23 @@
     }
   };
 
+  const mobileNavSection = (sectionIndex: number) => {
+    scrollToSection(sectionIndex);
+    closeMobileMenu();
+  };
+
+  const openMobileMenu = () => (isMobileMenuOpen.value = true);
+  const closeMobileMenu = () => (isMobileMenuOpen.value = false);
+
   watch(currentSectionIndex, (newValue: number) => {
     let previousSection = sections[newValue - 1]?.ref.value;
     let nextSection = sections[newValue + 1]?.ref.value;
 
-    if (!(previousSection?.tagName.toLowerCase() == 'section')) {
+    if (previousSection?.tagName.toLowerCase() !== 'section') {
       previousSection = previousSection?.querySelector('section');
     }
 
-    if (!(nextSection?.tagName.toLowerCase() == 'section')) {
+    if (nextSection?.tagName.toLowerCase() !== 'section') {
       nextSection = nextSection?.querySelector('section');
     }
 
@@ -504,31 +310,18 @@
   onMounted(() => {
     scrollToSection(sections.value[0].index);
 
-    setupProjectsList();
     setupInitialAnimations();
-
-    clipboardButton.value.addEventListener('click', () => {
-      if (clipboardShowCheck.value) return;
-
-      navigator.clipboard.writeText(MY_EMAIL).then(() => {
-        clipboardShowCheck.value = true;
-
-        snackbarStore.showSnackbar(
-          t('features.portfolio.sections.contact.email_copied_message'),
-        );
-
-        setTimeout(() => {
-          clipboardShowCheck.value = false;
-        }, 3000);
-      });
-    });
 
     window.addEventListener('wheel', (event: WheelEvent) =>
       handleMainScroll(event),
     );
+    content.value.addEventListener(
+      'scroll',
+      () => (showMobileHeader.value = content.value.scrollTop >= 400),
+    );
   });
 
-  interface ISection {
+  export interface ISection {
     id: string;
     index: number;
     title: string;
@@ -536,7 +329,7 @@
     navBarHref: string;
   }
 
-  interface ISocialLink {
+  export interface ISocialLink {
     label: string;
     url: string;
     icon: Component;
@@ -544,14 +337,124 @@
 </script>
 
 <style lang="scss" scoped>
+  .mobile-menu-container {
+    position: fixed;
+    top: 0;
+    left: 0;
+    height: 100vh;
+    width: 100vw;
+    overflow: hidden;
+    background-color: #0d0f12ac;
+    backdrop-filter: blur(16px);
+    z-index: 20;
+    display: flex;
+    justify-content: space-between;
+    flex-direction: column;
+    align-items: stretch;
+    transform: translateX(0);
+    transition: 0.7s;
+
+    &.hidden {
+      transform: translateX(100%);
+    }
+
+    .close-button {
+      display: flex;
+      padding: 24px;
+      justify-content: end;
+    }
+
+    .navigation {
+      flex-grow: 1;
+      display: flex;
+      padding: 40px;
+      align-items: center;
+      flex-direction: column;
+
+      .logo {
+        color: $primary-color;
+        transition: 0.4s scale, 0.8s color;
+        font-size: 32px;
+        text-decoration: none;
+        user-select: none;
+        margin-bottom: 64px;
+      }
+
+      ul {
+        padding: 12px 0;
+        gap: 20px;
+        display: flex;
+        list-style: none;
+        flex-direction: column;
+
+        li {
+          text-align: center;
+          width: 100%;
+          position: relative;
+
+          &:focus {
+            color: $secondary-color;
+          }
+
+          &.active {
+            .nav-item::after {
+              width: 50%;
+              left: 0;
+            }
+          }
+
+          .nav-item {
+            font-size: 16px;
+            font-weight: 300;
+            text-transform: uppercase;
+            text-decoration: none;
+            color: white;
+            position: relative;
+            box-sizing: content-box;
+            width: 100%;
+            transition: 0.5s;
+
+            &::after {
+              content: '';
+              position: absolute;
+              width: 0;
+              height: 2px;
+              border-radius: 8px;
+              background-color: $primary-color;
+              bottom: -4px;
+              left: 50%;
+              transition: 0.5s;
+            }
+          }
+        }
+      }
+    }
+
+    .social-links-container {
+      display: flex;
+      justify-content: center;
+      padding: 32px;
+    }
+  }
+
   .main-container {
     width: 100%;
-    min-height: 100vh;
+    height: 100%;
     position: relative;
-    padding: 0 4px;
-    overflow: hidden;
     display: flex;
+    overflow-x: hidden;
+    overflow-y: hidden;
     background-color: $background-color;
+
+    @media (width <= $desktop-min-width) {
+      overflow-y: auto;
+
+      .content {
+        section {
+          border-bottom: 1px solid rgba(255, 255, 255, 0.05);
+        }
+      }
+    }
 
     > :not(.content) {
       animation: show-element 1s 3s forwards;
@@ -568,6 +471,10 @@
       right: 0;
       opacity: 0;
       z-index: 10;
+
+      @media (width <= $desktop-min-width) {
+        display: none;
+      }
 
       .options {
         padding: 24px 0;
@@ -622,56 +529,42 @@
       transform: translateY(0);
       transition: transform 1s;
 
+      @media (width <= $desktop-min-width) {
+        display: none;
+      }
+
       &.hidden {
         transform: translateY(100%);
       }
+    }
 
-      .social-info {
+    .mobile-menu-button {
+      position: fixed;
+      top: 0;
+      left: 0;
+      width: 100vw;
+      background-color: #0d0f125e;
+      backdrop-filter: blur(16px);
+      z-index: 20;
+      display: none;
+      justify-content: flex-end;
+      padding: 16px 12px;
+      transition: transform 0.6s;
+
+      &.hidden {
+        transform: translateY(-100%);
+      }
+
+      @media (width <= $desktop-min-width) {
         display: flex;
-        bottom: 0;
-        flex-direction: column;
-        align-items: center;
-        gap: 24px;
-        justify-self: flex-end;
-
-        .social-link {
-          width: 24px;
-          height: 24px;
-          position: relative;
-          transition: 0.5s;
-
-          svg {
-            transform: scale(1.2, 1.2);
-            width: 24px;
-            height: 24px;
-            fill: white;
-            transition: 0.4s;
-          }
-
-          &:hover {
-            cursor: pointer;
-
-            svg,
-            svg path {
-              fill: $primary-color;
-            }
-          }
-        }
-
-        .separator {
-          height: 64px;
-          width: 1px;
-          background-color: white;
-        }
       }
     }
 
     .content {
       display: flex;
       flex-direction: column;
-      justify-self: center;
-      align-items: center;
       width: 100%;
+      height: 100%;
 
       .scroll-tip-container {
         opacity: 0;
@@ -700,9 +593,14 @@
         bottom: 0;
         opacity: 0;
         position: fixed;
+        left: 50%;
         justify-content: center;
         align-items: center;
         animation: show-element 1s 3s forwards;
+
+        @media (width <= $desktop-min-width) {
+          display: none;
+        }
 
         .next-section-container {
           display: flex;
@@ -750,536 +648,17 @@
         &.present {
           opacity: 0;
         }
-
-        &:first-child {
-          max-height: 80vh;
-        }
-
-        &#intro {
-          position: relative;
-          display: flex;
-          justify-content: space-between;
-          align-items: center;
-          gap: 12px;
-          animation: present-intro 3s forwards;
-
-          .profile-container {
-            display: flex;
-            justify-content: space-between;
-            align-items: center;
-
-            .profile {
-              display: flex;
-              flex-direction: column;
-              margin-left: 100px;
-              position: relative;
-
-              .name-container {
-                display: flex;
-                overflow: visible;
-
-                &.presented {
-                  .jp-name {
-                    &::before {
-                      animation: glitch-more 0.3s
-                        cubic-bezier(0.25, 0.46, 0.45, 0.94) 1;
-                    }
-
-                    &:after {
-                      animation: glitch-more 0.3s
-                        cubic-bezier(0.25, 0.46, 0.45, 0.94) reverse 1;
-                    }
-                  }
-                }
-
-                &:hover {
-                  .jp-name {
-                    &::before {
-                      animation: glitch 0.3s
-                        cubic-bezier(0.25, 0.46, 0.45, 0.94) both infinite;
-                    }
-
-                    &:after {
-                      animation: glitch 0.3s
-                        cubic-bezier(0.25, 0.46, 0.45, 0.94) reverse both
-                        infinite;
-                    }
-                  }
-                }
-
-                .pt-name {
-                  width: 460px;
-                  font-size: 64px;
-                  line-height: 64px;
-                  font-weight: 900;
-                  text-transform: uppercase;
-                }
-
-                .jp-name {
-                  color: $primary-color;
-                  font-size: 100px;
-                  position: absolute;
-                  z-index: -1;
-                  top: 0;
-                  right: 35%;
-                  opacity: 0;
-                  transform: translate(-20px, -25%);
-                  animation: show-element 1s 2s forwards;
-
-                  &:before,
-                  &:after {
-                    display: block;
-                    content: attr(data-text);
-                    position: absolute;
-                    top: 0;
-                    left: 0;
-                    opacity: 0.8;
-                  }
-
-                  &:before {
-                    color: $primary-color;
-                    z-index: -1;
-                  }
-
-                  &:after {
-                    color: $secondary-color;
-                    z-index: -2;
-                  }
-                }
-              }
-
-              .role {
-                font-weight: 200;
-                font-size: 40px;
-              }
-
-              .cv-link {
-                text-decoration: none;
-                display: flex;
-                justify-content: center;
-                align-items: center;
-                background-color: $background-color;
-                color: white;
-                font-weight: 200;
-                cursor: pointer;
-                align-self: flex-start;
-                transition: 0.3s;
-                text-transform: uppercase;
-                padding: 24px 40px;
-                font-size: 16px;
-                border: 1px solid rgba(255, 255, 255, 0.3);
-                line-height: 20px;
-                margin-top: 24px;
-
-                &:hover {
-                  font-weight: 500;
-                  border: 1px solid transparent;
-                  background-color: $primary-color;
-                  color: $background-color;
-                }
-              }
-            }
-
-            .photo-container {
-              width: 400px;
-              height: 400px;
-              background-size: cover;
-              margin-right: 100px;
-              transition: 0.3s;
-              box-shadow: 2px 2px 1px 1px $background-color,
-                56px 56px 1px 1px $primary-color;
-
-              &:hover {
-                box-shadow: 2px 2px 1px 1px $background-color,
-                  24px 24px 1px 1px $primary-color;
-              }
-            }
-          }
-        }
-
-        &#projects {
-          display: flex;
-          flex-direction: column;
-          padding: 100px;
-          gap: 12px;
-
-          .description {
-            width: 50%;
-            font-size: 20px;
-            font-weight: 300;
-            margin-bottom: 32px;
-          }
-
-          .projects-container {
-            display: flex;
-            width: 100%;
-            height: 100%;
-            position: relative;
-
-            &:hover {
-              .previous-button,
-              .next-button {
-                opacity: 1;
-                transform: translate(0, -100%);
-              }
-            }
-
-            .previous-button,
-            .next-button {
-              position: absolute;
-              display: flex;
-              opacity: 0;
-              justify-content: center;
-              align-items: center;
-              backdrop-filter: blur(4px);
-              top: 50%;
-              padding: 12px;
-              border-radius: 50px;
-              transition: 0.5s, opacity 0.2s;
-              background-color: #0d0f1279;
-              border: 1px solid rgba(255, 255, 255, 0.7);
-              cursor: pointer;
-
-              &:hover {
-                backdrop-filter: blur(8px);
-                background-color: #0d0f12b8;
-              }
-
-              &.hidden {
-                cursor: default;
-                opacity: 0;
-              }
-            }
-
-            .previous-button {
-              left: 0;
-              transform: translate(50px, -100%);
-            }
-
-            .next-button {
-              right: 0;
-              transform: translate(-50px, -100%);
-            }
-
-            .projects-list {
-              display: flex;
-              gap: 24px;
-              overflow-x: auto;
-              width: 100%;
-              height: 100%;
-              padding-bottom: 16px;
-              padding-right: 100px;
-
-              mask-image: linear-gradient(
-                to left,
-                transparent,
-                $background-color 1%,
-                $background-color 99%,
-                transparent
-              );
-              -webkit-mask-image: linear-gradient(
-                to left,
-                transparent,
-                $background-color 1%,
-                $background-color 99%,
-                transparent
-              );
-            }
-          }
-        }
-
-        &#contact {
-          height: 100vh;
-          display: flex;
-          flex-direction: column;
-          align-items: stretch;
-          justify-content: space-between;
-          padding: 0;
-
-          .contact-container {
-            padding: 120px;
-            gap: 12px;
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-
-            .description {
-              font-size: 20px;
-              font-weight: 400;
-            }
-
-            .email-contact-container {
-              display: flex;
-              gap: 12px;
-              align-items: center;
-
-              .email-contact {
-                color: white;
-              }
-
-              .clipboard-button {
-                width: 24px;
-                height: 24px;
-
-                &:hover {
-                  cursor: pointer;
-
-                  .clipboard-plain {
-                    &:deep(rect:first-child) {
-                      stroke: $primary-color !important;
-                      fill: $primary-color !important;
-                    }
-
-                    &:deep(rect:last-child) {
-                      stroke: $primary-color !important;
-                    }
-                  }
-                }
-              }
-            }
-          }
-
-          footer {
-            display: flex;
-            align-items: center;
-            justify-content: space-evenly;
-            justify-self: stretch;
-            border-top: 1px solid #ffffff25;
-            gap: 12px;
-
-            > div,
-            > span {
-              flex: 1 0;
-              display: flex;
-              justify-content: center;
-            }
-
-            .social-links-container {
-              flex: 1 0;
-
-              .social-link svg {
-                width: 20px;
-                height: 20px;
-              }
-            }
-
-            span,
-            a {
-              font-size: 12px;
-            }
-
-            .repo-tip {
-              display: flex;
-              gap: 4px;
-              font-size: 12px;
-
-              a {
-                color: $secondary-color;
-              }
-            }
-          }
-        }
       }
-    }
-
-    .navbar-container {
-      position: fixed;
-      width: 100%;
-      z-index: 10;
-      animation: show-element 1s forwards;
-      transform: translateY(0);
-      transition: 0.5s;
-      display: flex;
-      flex-direction: column-reverse;
-      align-items: center;
-
-      &.hidden {
-        transform: translateY(-100%);
-      }
-
-      .vertical-line {
-        width: 1px;
-        height: 24px;
-        background-color: white;
-      }
-
-      .horizontal-line {
-        width: 10%;
-        height: 1px;
-        transition: 1s;
-        background-color: white;
-      }
-
-      .nav-bar {
-        z-index: 5;
-        display: flex;
-        width: 100%;
-        justify-content: space-between;
-        padding: 0 32px;
-
-        &:hover {
-          ~ .horizontal-line {
-            width: 40%;
-          }
-        }
-
-        .logo {
-          display: flex;
-          justify-content: center;
-          align-items: center;
-          color: white;
-          transition: 0.4s scale, 0.8s color;
-          font-size: 24px;
-          text-decoration: none;
-          user-select: none;
-          position: relative;
-          animation: show-element 1s forwards;
-
-          &:last-child {
-            color: transparent;
-            pointer-events: none;
-          }
-
-          &:before,
-          &:after {
-            width: 100%;
-            display: block;
-            font-size: 24px;
-            content: attr(data-text);
-            position: absolute;
-            top: 50%;
-            left: 50%;
-            transform: translate(-50%, -50%);
-            opacity: 0.8;
-          }
-
-          &::before {
-            color: $primary-color !important;
-            z-index: -1;
-          }
-
-          &::before {
-            color: $primary-color !important;
-            z-index: -1;
-          }
-
-          &:hover {
-            color: $primary-color;
-            scale: 1.1;
-
-            &::before {
-              animation: glitch-logo 0.3s cubic-bezier(0.25, 0.46, 0.45, 0.94)
-                both 1;
-            }
-
-            &:after {
-              animation: glitch-logo 0.3s cubic-bezier(0.25, 0.46, 0.45, 0.94)
-                both 1;
-            }
-          }
-        }
-
-        ul {
-          padding: 12px 0;
-          gap: 20px;
-          display: flex;
-          list-style: none;
-
-          li {
-            width: 100%;
-            position: relative;
-
-            &:hover {
-              color: $secondary-color;
-
-              .nav-item::before {
-                opacity: 1;
-                transform: translate(-12px, -50%);
-              }
-
-              .nav-item::after {
-                opacity: 1;
-                transform: translate(12px, -50%);
-              }
-            }
-
-            &::after {
-              content: '';
-              position: absolute;
-              width: 0;
-              height: 2px;
-              border-radius: 8px;
-              background-color: $primary-color;
-              bottom: 0;
-              left: 0;
-              transition: 0.5s;
-            }
-
-            &.active {
-              .nav-item {
-                font-weight: 500;
-              }
-
-              &::after {
-                width: 50%;
-              }
-
-              &:hover {
-                &::after {
-                  width: 0;
-                }
-              }
-            }
-
-            .nav-item {
-              font-size: 16px;
-              font-weight: 300;
-              text-transform: uppercase;
-              text-decoration: none;
-              color: white;
-              position: relative;
-              box-sizing: content-box;
-              width: 100%;
-              transition: 0.5s;
-
-              &::before {
-                content: '{';
-                position: absolute;
-                color: $primary-color;
-                font-size: 20px;
-                opacity: 0;
-                top: 50%;
-                left: 0;
-                transform: translateY(-50%);
-                transition: 0.4s;
-              }
-
-              &::after {
-                content: '}';
-                position: absolute;
-                color: $primary-color;
-                font-size: 20px;
-                opacity: 0;
-                top: 50%;
-                right: 0;
-                transform: translateY(-50%);
-                transition: 0.4s;
-              }
-            }
-          }
-        }
-      }
-    }
-
-    .content {
-      padding: 80px 40px;
     }
   }
 
-  .social-links-container {
+  .social-info {
     display: flex;
-    gap: 12px;
-    margin: 12px 0;
+    bottom: 0;
+    flex-direction: column;
+    align-items: center;
+    gap: 24px;
+    justify-self: flex-end;
 
     .social-link {
       width: 24px;
@@ -1304,105 +683,11 @@
         }
       }
     }
-  }
 
-  @keyframes present-intro {
-    0% {
-      transform: translateY(-100%);
-      opacity: 0;
-    }
-    50% {
-      transform: translateY(0);
-      opacity: 0.5;
-    }
-    100% {
-      transform: translateY(0);
-      opacity: 1;
-    }
-  }
-
-  @keyframes show-element {
-    from {
-      opacity: 0;
-    }
-
-    to {
-      opacity: 1;
-    }
-  }
-
-  @keyframes shimmer {
-    to {
-      mask-position: 0 250%;
-    }
-  }
-
-  @keyframes glitch {
-    0% {
-      transform: translate(0);
-    }
-    20% {
-      transform: translate(-2px, 2px);
-      z-index: 2;
-    }
-    40% {
-      transform: translate(-2px, -2px);
-    }
-    60% {
-      transform: translate(2px, 2px);
-      z-index: 1;
-    }
-    80% {
-      transform: translate(2px, -2px);
-    }
-    to {
-      transform: translate(0);
-    }
-  }
-
-  @keyframes glitch-logo {
-    0% {
-      transform: translate(-50%, -50%);
-    }
-    20% {
-      transform: translate(-40%, -60%);
-      z-index: 2;
-    }
-    40% {
-      transform: translate(-40%, -40%);
-    }
-    60% {
-      transform: translate(-60%, -60%);
-      z-index: 1;
-    }
-    80% {
-      transform: translate(-51%, -40%);
-    }
-    to {
-      transform: translate(-50%, -50%);
-    }
-  }
-
-  @keyframes glitch-more {
-    0% {
-      transform: translate(0);
-    }
-    20% {
-      transform: translate(-6px, 4px);
-      z-index: 2;
-    }
-    40% {
-      transform: translate(-4px, -4px);
-    }
-    60% {
-      transform: translate(4px, 6px);
-      z-index: 1;
-    }
-    80% {
-      transform: translate(4px, -4px);
-    }
-    to {
-      transform: translate(0);
+    .separator {
+      height: 64px;
+      width: 1px;
+      background-color: white;
     }
   }
 </style>
